Add worker-only start support
Make the node a worker node until user issues microk8s start.
New process:

  $ sudo snap install --classic microk8s
  # microk8s installed and partially configured to be a worker node
  # Undo it:
  $ microk8s start --all
  # Or keep the node as worker:
  $ microk8s start

Join a cluster as a worker node with:

  $ microk8s join --worker

That saves RAM in the initial start process by avoiding the API
processes to come up.

The microk8s start command will now retrigger the configure hook
by running a "snap set" before the "snap start". This ensures that
new configuration files to be generate in case the user is promoting
this node from worker to master.